requirements.txt
spotipy>=2.19.0
python-telegram-bot>=20.0
pandas>=1.5.0
scikit-learn>=1.2.0
python-dotenv>=0.21.0
scipy>=1.10.0
matplotlib>=3.7.0
openai>=1.0.0
